lumea-api/requirements.txt
fastapi
uvicorn[standard]
groq
httpx
python-dotenv
pydantic
